{"schema":"decodex/commit/1","summary":"Modularize permissions settings panel","authority":"manual"}
