+++ docs/reference/host-core-reset.md
@@ -164,9 +164,10 @@ The current native-host Swift split is:
   models, `FrozenOverlayFFI.swift` owns frozen overlay edit/export bridge models, and
   `HostFFISupport.swift` owns shared status, geometry, and owned-buffer adaptation helpers.
 - `NativeHostSettingsView.swift`, `NativeHostSettingsNavigation.swift`,
-  `NativeHostSettingsSurface.swift`, `NativeHostAppearanceSettings.swift`, and
-  `NativeHostCaptureFrameSettings.swift`: SwiftUI settings shell, navigation, reusable settings
-  surfaces, appearance controls, and capture-frame preset controls.
+  `NativeHostSettingsSurface.swift`, `NativeHostAppearanceSettings.swift`,
+  `NativeHostCaptureFrameSettings.swift`, and `PermissionsSettingsPanel.swift`: SwiftUI settings
+  shell, navigation, reusable settings surfaces, appearance controls, capture-frame preset
+  controls, and permission/setup controls.
 
 This means a large Swift line count can still be reasonable when those lines are AppKit,
 CoreGraphics, ScreenCaptureKit, Vision, pasteboard, save-panel, sound, update, or window lifecycle
